Fix issue with plotting fluorescence and ADC channel of different lengths.
In the Time Course Analysis window, Jaime Guzman pointed out that if the
fluorescence and voltage stimulus traces in an .IDR file were different
lengths, this would cause issues with plotting (the trace would be truncated,
and the values of the time axis would be off by a factor, in the file we
looked at it was 1.5).  The maximum of the two lengths is now used.