Allow existing waveforms to be dragged and dropped in stimulus editor.
Previously a new waveform could be dragged and dropped onto an existing
one, effectively erasing the previous waveform, replacing it with a
"blank" waveform of the same or a different type.  This change allows an
existing waveform to be dragged and dropped onto another; the parameters
of the second are overwritten by those of the first.