Revert change of "A/D Signals Only" in Record Signals window clicked by default
Change was made because of suggestion that photostimulus workflow would
be improved without having to select this each time.  But others reported
that this was detrimental to linescan workflow, in which case user would
need to keep unchecking this box.  Box is now unchecked by default as it
was originally, but retains state between closing and reopening Record
Signals window.